flask/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pigpio==1.78
RPi.GPIO==0.7.1
smbus2==0.5.0
Werkzeug==3.1.4
